Pattern - Changed the pattern to work for realistic use cases
now the response ADT is not driven by how many states can the machine be in. Instead, it is driven by how what messages it can serve. For example, instead of just two cases: READ & WRITE, now we have 3 cases: read, clear and save. So now, responses are related to messages and not states.